Chat #8 Debug
Okay your code is good but there are two bugs. First of all, the main function should be:

    app = QApplication(sys.argv)
    model = Model()
    view = View()
    controller = Controller(model, view)
    controller.view.show()
    sys.exit(app.exec_())

If you don't understand this, try thinking about the GC mechanism in Python.

Secondly, the model should NOT be passed to the view since they should NOT communicate.

I already fixed the bug so you DO NOT have to output the code for me. I just wanted you to be aware of the issues